Fix 'signed overflow' related build warning
Fix the following build warning:

/builddir/build/BUILD/bpftrace-master/src/bpftrace.cpp: In member
function 'int bpftrace::BPFtrace::print_lhist(const std::vector<long
unsigned int>&, int, int, int) const':
/builddir/build/BUILD/bpftrace-master/src/bpftrace.cpp:1880:1: warning:
assuming signed overflow does not occur when changing X +- C1 cmp C2 to
X cmp C2 -+ C1 [-Wstrict-overflow]
 1880 | } // namespace bpftrace
      | ^

+++ src/bpftrace.cpp
@@ -1283,7 +1283,7 @@ int BPFtrace::print_lhist(const std::vector<uint64_t> &values, int min, int max,
   int start_value = -1;
   int end_value = 0;
 
-  for (int i = 0; i <= buckets + 1; i++)
+  for (unsigned int i = 0; i <= static_cast<unsigned int>(buckets) + 1; i++)
   {
     if (values.at(i) > 0) {
       if (start_value == -1) {
